src/requirements.txt
httpx
pydantic
streamlit
python-dotenv
